+++ servers/quarkus-server/src/test/java/org/projectnessie/server/TestQuarkusEventsEnabledAuthEnabled.java
@@ -28,10 +28,8 @@ import org.projectnessie.client.NessieClientBuilder;
 import org.projectnessie.client.auth.BasicAuthenticationProvider;
 import org.projectnessie.client.ext.NessieApiVersion;
 import org.projectnessie.client.ext.NessieApiVersions;
-import org.projectnessie.events.service.EventSubscribers;
 import org.projectnessie.server.authn.AuthenticationEnabledProfile;
 import org.projectnessie.server.events.EventsEnabledProfile;
-import org.projectnessie.server.events.fixtures.MockEventSubscriber;
 
 @QuarkusTest
 @TestProfile(TestQuarkusEventsEnabledAuthEnabled.Profile.class)
@@ -53,7 +51,6 @@ public class TestQuarkusEventsEnabledAuthEnabled extends AbstractQuarkusEvents {
     }
   }
 
-  @Inject Instance<EventSubscribers> subscribers;
   @Inject Instance<MeterRegistry> registries;
 
   @Override
@@ -70,11 +67,6 @@ public class TestQuarkusEventsEnabledAuthEnabled extends AbstractQuarkusEvents {
   @Override
   protected String eventInitiator() {
     return "test_user";
-  }
-
-  @Override
-  protected MockEventSubscriber subscriber() {
-    return (MockEventSubscriber) subscribers.get().getSubscribers().get(0);
   }
 
   @Override
